requirements.txt
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.20
pandas
sqlalchemy==2.0.25
streamlit==1.30.0
plotly==5.18.0
python-dotenv==1.0.0
pytest==7.4.4
fpdf==1.7.2
pypdf==3.17.4
chromadb==0.4.22
